+++ posts/data/xml_post_data.py
@@ -1,20 +1,14 @@
 from xml.etree import ElementTree
 
-from discord import Embed
-
 from posts.data.post_data import PostData
 
 
 class XmlPostData(PostData):
-    def __init__(self, total_posts=0, **kwargs):
-        super().__init__(**kwargs)
-        self.total_posts: int = total_posts
-
     @classmethod
-    def from_xml(cls, el: ElementTree, total_posts: int):
+    def from_xml(cls, el: ElementTree):
         file_url = el.get('file_url')
 
-        data = dict(
+        return cls(
             file_url=file_url,
             file_ext=file_url.split('.')[-1],
             created_at=el.get('created_at'),
@@ -23,10 +17,3 @@ class XmlPostData(PostData):
             tags=el.get('tags'),
             id=el.get('id')
         )
-
-        return cls(total_posts, **data)
-
-    def to_embed(self) -> Embed:
-        embed = super().to_embed()
-        embed.description = f'Found {self.total_posts} images'
-        return embed
