+++ posts/post/xml_post.py
@@ -5,13 +5,13 @@ from xml.etree import ElementTree
 from discord.ext.commands import Context, CommandError
 
 from posts.api.xml_api import send_request
-from posts.data.post_data import PostHasDisallowedTags, PostData
+from posts.data.post_data import PostData
 from posts.data.xml_post_data import XmlPostData
-from posts.post.post import AbstractPost
+from posts.post.post import Post
 from util import util
 
 
-class XmlPost(AbstractPost):
+class XmlPost(Post):
     total_posts: int = 0
 
     def __init__(self, ctx: Context, url: str, tags: str):
@@ -26,14 +26,10 @@ class XmlPost(AbstractPost):
         await super().create_message()
 
     def fetch_post(self) -> PostData:
-        xml_post = random_xml_post(self.url, self.tags, self.total_posts)
+        random_page = random.randint(0, self.total_posts - 1)
+        xml_post = fetch_xml_post(self.url, self.tags, random_page)
 
-        post_data = XmlPostData.from_xml(xml_post, self.total_posts)
-        if post_data.has_disallowed_tags():
-            return PostHasDisallowedTags()
-
-        self.update_hist(post_data)
-        return post_data
+        return xml_post
 
 
 async def show_post(ctx: Context, tags: str, score: int, url: str, skip_score=False):
@@ -45,13 +41,7 @@ async def show_post(ctx: Context, tags: str, score: int, url: str, skip_score=Fa
 
 
 def get_total_posts(url: str, tags: str) -> int:
-    """
-    Returns the total amount of posts for a list of tags
-
-    :param url: the url to search the pots for
-    :param tags: the tags
-    :return: the total amount of posts for a tag
-    """
+    # Fetch 0 posts to just get the post count
     resp_text = send_request(url, 0, tags, 0)
     posts = ElementTree.fromstring(resp_text)
 
@@ -63,10 +53,11 @@ def get_total_posts(url: str, tags: str) -> int:
     return 0
 
 
-def random_xml_post(url: str, tags: str, total_posts: int) -> Element:
-    random_page = random.randint(0, total_posts - 1)
-
-    resp_text = send_request(url, 1, tags, random_page)
+def fetch_xml_post(url: str, tags: str, page: int) -> Element:
+    resp_text = send_request(url, 1, tags, page)
     posts = ElementTree.fromstring(resp_text)
 
-    return posts[0]
+    if len(posts) == 0:
+        return
+
+    return XmlPostData.from_xml(posts[0])
